refactor(cli): split ai.py command monolith into a package
Break the 2601-line openfatture/cli/commands/ai.py into a package: _app (main
Typer + 4 sub-Typers) + one module per command group (rag, feedback, retrain,
auto_update, self_learning, describe, vat, forecast, compliance, create_invoice,
chat, voice). Sub-apps mounted in __init__ with identical names; command bodies
are pure copy-paste.

Tests repointed (no shim): mock.patch targets moved from commands.ai.<symbol>
to the module owning each exercised command; shared seams (create_provider,
enrich_with_rag) repointed per the command each test invokes.

Verified zero regression: command tree identical (23 entries, golden diff empty);
the 3 affected test files yield the EXACT same pass/fail set as the original
monolith (45 passed / 16 failed — the 16 are pre-existing, environment/DB-bound
failures unrelated to this split, confirmed by running both states). `openfatture
ai --help` OK, ruff clean.
